LPS-60850 Move the base to API and create the implementations as components
