+++ modules/apps/collaboration/wiki/wiki-api/bnd.bnd
@@ -14,6 +14,7 @@ Export-Package:\
 	com.liferay.wiki.model,\
 	com.liferay.wiki.service,\
 	com.liferay.wiki.service.persistence,\
+	com.liferay.wiki.translator,\
 	com.liferay.wiki.validator
 Liferay-Releng-Module-Group-Description:
 Liferay-Releng-Module-Group-Title: Wiki
