+++ modules/apps/collaboration/wiki/wiki-api/src/main/java/com/liferay/wiki/translator/BaseTranslator.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.wiki.internal.translator;
+package com.liferay.wiki.translator;
 
 import com.liferay.portal.kernel.util.DigesterUtil;
 import com.liferay.portal.kernel.util.StringPool;
