+++ modules/apps/collaboration/wiki/wiki-service/bnd.bnd
@@ -11,6 +11,7 @@ Export-Package:\
 	com.liferay.wiki.service.permission,\
 	com.liferay.wiki.service.util,\
 	com.liferay.wiki.social,\
+	com.liferay.wiki.translator.impl,\
 	com.liferay.wiki.trash,\
 	com.liferay.wiki.util,\
 	com.liferay.wiki.util.comparator,\
