+++ modules/apps/collaboration/wiki/wiki-service/build.gradle
@@ -15,7 +15,6 @@ dependencies {
 	provided group: "com.liferay", name: "com.liferay.portlet.display.template", version: "2.0.0"
 	provided group: "com.liferay", name: "com.liferay.registry.api", version: "1.0.0"
 	provided group: "com.liferay", name: "com.liferay.rss.util", version: "1.0.0"
-	provided group: "com.liferay", name: "com.liferay.wiki.api", version: "2.1.0"
 	provided group: "com.liferay", name: "com.liferay.xstream.configurator.api", version: "2.0.0"
 	provided group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "2.0.0"
 	provided group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "2.6.0"
@@ -28,4 +27,5 @@ dependencies {
 	provided group: "org.osgi", name: "org.osgi.core", version: "5.0.0"
 	provided group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
 	provided group: "rome", name: "rome", version: "1.0"
+	provided project(":apps:collaboration:wiki:wiki-api")
 }
