+++ modules/apps/collaboration/wiki/wiki-test/build.gradle
@@ -1,8 +1,6 @@
 dependencies {
 	testCompile group: "com.liferay", name: "com.liferay.exportimport.api", version: "2.0.0"
 	testCompile group: "com.liferay", name: "com.liferay.exportimport.service", version: "3.0.0"
-	testCompile group: "com.liferay", name: "com.liferay.wiki.api", version: "2.0.0"
-	testCompile group: "com.liferay", name: "com.liferay.wiki.service", version: "1.0.0"
 	testCompile group: "com.liferay", name: "com.liferay.wiki.web", version: "1.0.0"
 	testCompile group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "2.0.0"
 
@@ -10,4 +8,7 @@ dependencies {
 	testIntegrationCompile group: "com.liferay", name: "com.liferay.registry.api", version: "1.0.0"
 	testIntegrationCompile group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "2.0.0"
 	testIntegrationCompile group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "2.0.0"
+
+	testIntegrationCompile project(":apps:collaboration:wiki:wiki-api")
+	testIntegrationCompile project(":apps:collaboration:wiki:wiki-service")
 }
